Update login page
