+++ node/src/views/contacts.pug
@@ -1,4 +1,8 @@
 extends includes/layout
+
+block scripts
+  script(type='application/javascript' src='/js/contacts.js')
+
 block content
   .container
     h3(class='alert-not-shown-events') New webhooks are received.
