+++ node/src/views/includes/footer.pug
@@ -1,2 +1,3 @@
+block footerScripts
 .footer
   .container
