+++ node/src/views/includes/head.pug
@@ -9,6 +9,5 @@ head
   // Fav Icon
   link(href='/favicon-32x32.webp' rel='shortcut icon')
   link(href='/favicon-32x32.webp' rel='apple-touch-icon')
-
+block scripts
   script(type='application/javascript' src='https://cdnjs.cloudflare.com/ajax/libs/jquery/3.4.1/jquery.min.js')
-  script(type='application/javascript' src='/js/main.js')
